+++ src/random.h
@@ -7,7 +7,7 @@ typedef std::mt19937 generator_t;
 
 void init_random();
 int get_random_int(int, int);
-double get_random_float(double, double);
-double get_random_prob();
+float get_random_float(float, float);
+float get_random_prob();
 
 #endif
